backend/requirements/base.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg[binary]>=3.1,<4.0
python-dotenv>=1.0,<2.0
whitenoise>=6.7,<7.0
